v3.0.0
- MAJOR OVERHAUL: This plugin has a major overhaul and is completely rewritten!
Now you've the ability to choose witch items are sold or bought, from a specific
store. You can create your own stores with custom sold items, and the stores can
be configured from the config. For this overhaul we've stopped using the database
for storing store information, all API calls to this database table are now
deprecated. The sell page of the store only shows the items the NPC will sell if
you have these items in your inventory. When your inventory is empty you do not
see what you can sell at this NPC. A video how this plugin works will be released
soon.

- Added: Ability to quickly set up shops.
- Added: Ability to sync the items between all shops with the same type id.
- Added: Ability to customize items with a synced store to make them unique.
- Added: All butchers - general stores on the official rdr2 locations
- Added: The bait shop on the RDR2 default location.
- Added: Carpenter, Herbs, Mineral, Vegetables, and Gold Hoarders shops
- Added: Option to show or disable the switch button in the store menu.